protect against failed eval of setup.py code fragments
When we're parsing setup.py instead of running it, we extract code
fragments, and then eval them in the hope of obtaining the correct
value.  However this is highly prone to failure, so at least attempt to
fail gracefully in a way which guides the user towards the right manual
fix.

This does not fix https://github.com/saschpe/py2pack/issues/41 but
at least it makes it clearer what the issue is.