warn when --run fails to retrieve dist object (#41)
It seems that --run is no longer working reliably, even if we
make the following change to _run_setup_py():

-    setuptools.sandbox.run_setup(setup_filename, "")
+    setuptools.sandbox.run_setup(setup_filename, ["check"])

So check for when we fail to get a dist object back from which
we can retrieve values, and display the failure gracefully to
the user.

This does not fix https://github.com/saschpe/py2pack/issues/41 but
at least it makes it clearer what the issue is.

A far better approach to this problem of dynamic evaluation of
metadata from setup.py is shown by fpm:

https://github.com/jordansissel/fpm/blob/master/lib/fpm/package/python.rb
https://github.com/jordansissel/fpm/blob/master/lib/fpm/package/pyfpm/get_metadata.py

That approach does not rely on any private internal implementation
details within distutils.