Set setup.py encoding and retry parsing in error case
When calling "python setup.py --command-packages py2pack get_metadata"
fails, set as first line in the setup.py the encoding to UTF-8 according
to PEP-0263. Then retry with the set encoding.
This might be not correct because we don't know the encoding of the file
for sure but it's at least a try.