Add a --source-url parameter to 'fetch' and 'generate' commands
There are some packages in pypi that don't define a source url
(for example, SpeechRecognition, whose sources are only released
in github). In those cases, now it's possible to specify a
--source-url="http://github.com/whatever" parameter to fetch
the sources and generate the spec file automatically anyway.

Note that it's recommended to check the sources statement in
the generated spec file (for example, to replace the version
string with %{version})